c https://x.com/ubugeeei/status/2012573998249582883?s=20

+++ src/main/java/myau/module/modules/AutoBlockIn.java
@@ -30,7 +30,7 @@ public class AutoBlockIn extends Module {
     private static final Minecraft mc = Minecraft.getMinecraft();
     private final Map<String, Integer> BLOCK_SCORE = new HashMap<>();
     private long lastPlaceTime = 0;
-    
+
     public final FloatProperty range = new FloatProperty("range", 4.5f, 3.0f, 6.0f);
     public final IntProperty speed = new IntProperty("speed", 20, 5, 100);
     public final IntProperty placeDelay = new IntProperty("place-delay", 50, 0, 200);
@@ -38,7 +38,7 @@ public class AutoBlockIn extends Module {
     public final BooleanProperty itemSpoof = new BooleanProperty("item-spoof", true);
     public final BooleanProperty showProgress = new BooleanProperty("show-progress", true);
     public final ModeProperty moveFix = new ModeProperty("move-fix", 1, new String[]{"NONE", "SILENT", "STRICT"});
-    
+
     private float serverYaw;
     private float serverPitch;
     private float progress;
@@ -48,15 +48,15 @@ public class AutoBlockIn extends Module {
     private EnumFacing targetFacing;
     private Vec3 targetHitVec;
     private int lastSlot = -1;
-    
-    private static final int[][] DIRS = {{1,0,0}, {0,0,1}, {-1,0,0}, {0,0,-1}};
+
+    private static final int[][] DIRS = {{1, 0, 0}, {0, 0, 1}, {-1, 0, 0}, {0, 0, -1}};
     private static final double INSET = 0.05;
     private static final double STEP = 0.2;
     private static final double JIT = STEP * 0.1;
-    
+
     public AutoBlockIn() {
         super("AutoBlockIn", false);
-        
+
         BLOCK_SCORE.put("obsidian", 0);
         BLOCK_SCORE.put("end_stone", 1);
         BLOCK_SCORE.put("planks", 2);
@@ -100,14 +100,14 @@ public class AutoBlockIn extends Module {
         if (!isEnabled()) return;
         if (event.getType() != EventType.PRE) return;
         if (mc.thePlayer == null || mc.theWorld == null) return;
-        
+
         if (mc.currentScreen != null) {
             return;
         }
-        
+
         serverYaw = event.getYaw();
         serverPitch = event.getPitch();
-        
+
         updateProgress();
 
         int blockSlot = findBestBlockSlot();
@@ -117,7 +117,7 @@ public class AutoBlockIn extends Module {
                 mc.thePlayer.inventory.currentItem = blockSlot;
             }
         }
-        
+
         ItemStack currentHeld = mc.thePlayer.inventory.getCurrentItem();
         boolean holdingBlock = currentHeld != null && currentHeld.getItem() instanceof ItemBlock;
         if (!holdingBlock) {
@@ -126,31 +126,31 @@ public class AutoBlockIn extends Module {
             targetHitVec = null;
             return;
         }
-        
+
         findBestPlacement();
-        
+
         if (targetBlock != null && targetFacing != null && targetHitVec != null) {
             Vec3 eyes = mc.thePlayer.getPositionEyes(1.0f);
             double dx = targetHitVec.xCoord - eyes.xCoord;
             double dy = targetHitVec.yCoord - eyes.yCoord;
             double dz = targetHitVec.zCoord - eyes.zCoord;
             double dist = Math.sqrt(dx * dx + dz * dz);
-            
-            float targetYaw = (float)Math.toDegrees(Math.atan2(dz, dx)) - 90.0f;
-            float targetPitch = (float)-Math.toDegrees(Math.atan2(dy, dist));
-            
+
+            float targetYaw = (float) Math.toDegrees(Math.atan2(dz, dx)) - 90.0f;
+            float targetPitch = (float) -Math.toDegrees(Math.atan2(dy, dist));
+
             targetYaw = MathHelper.wrapAngleTo180_float(targetYaw);
-            
+
             float yawDiff = MathHelper.wrapAngleTo180_float(targetYaw - serverYaw);
             float pitchDiff = targetPitch - serverPitch;
-            
+
             float maxTurn = speed.getValue().floatValue();
             float yawStep = MathHelper.clamp_float(yawDiff, -maxTurn, maxTurn);
             float pitchStep = MathHelper.clamp_float(pitchDiff, -maxTurn, maxTurn);
-            
+
             aimYaw = serverYaw + yawStep;
             aimPitch = MathHelper.clamp_float(serverPitch + pitchStep, -90.0f, 90.0f);
-            
+
             event.setRotation(aimYaw, aimPitch, 6);
             event.setPervRotation(this.moveFix.getValue() != 0 ? aimYaw : mc.thePlayer.rotationYaw, 6);
         }
@@ -167,44 +167,44 @@ public class AutoBlockIn extends Module {
             }
         }
     }
-    
+
     @EventTarget(Priority.HIGH)
     public void onTick(TickEvent event) {
         if (!isEnabled()) return;
         if (event.getType() != EventType.PRE) return;
         if (mc.thePlayer == null || mc.theWorld == null) return;
-        
+
         if (mc.currentScreen != null) {
             return;
         }
-        
+
         if (targetBlock != null && targetFacing != null && targetHitVec != null) {
             if (!withinRotationTolerance(aimYaw, aimPitch)) {
                 return;
             }
-            
+
             long currentTime = System.currentTimeMillis();
             if (currentTime - lastPlaceTime >= placeDelay.getValue()) {
                 lastPlaceTime = currentTime;
-                
+
                 MovingObjectPosition mop = rayTraceBlock(aimYaw, aimPitch, range.getValue());
-                
-                if (mop != null 
+
+                if (mop != null
                         && mop.typeOfHit == MovingObjectPosition.MovingObjectType.BLOCK
                         && mop.getBlockPos().equals(targetBlock)
                         && mop.sideHit == targetFacing) {
-                    
+
                     ItemStack heldStack = mc.thePlayer.inventory.getCurrentItem();
                     if (heldStack != null && heldStack.getItem() instanceof ItemBlock) {
                         mc.playerController.onPlayerRightClick(
-                            mc.thePlayer,
-                            mc.theWorld,
-                            heldStack,
-                            targetBlock,
-                            targetFacing,
-                            mop.hitVec);
+                                mc.thePlayer,
+                                mc.theWorld,
+                                heldStack,
+                                targetBlock,
+                                targetFacing,
+                                mop.hitVec);
                         mc.thePlayer.swingItem();
-                        
+
                         targetBlock = null;
                         targetFacing = null;
                         targetHitVec = null;
@@ -221,35 +221,35 @@ public class AutoBlockIn extends Module {
             event.setCancelled(true);
         }
     }
-    
+
     @EventTarget
     public void onRender2D(Render2DEvent event) {
         if (!isEnabled() || mc.currentScreen != null) return;
         if (!showProgress.getValue()) return;
         if (mc.fontRendererObj == null) return;
-        
+
         float scale = 1.0f;
         String text = String.format("Blocking: %.0f%%", progress * 100.0F);
-        
+
         GL11.glPushMatrix();
-        GL11.glScaled((double)scale, (double)scale, 0.0);
+        GL11.glScaled(scale, scale, 0.0);
         GlStateManager.disableDepth();
         GlStateManager.enableBlend();
         GlStateManager.blendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA);
-        
+
         ScaledResolution sr = new ScaledResolution(mc);
         int width = mc.fontRendererObj.getStringWidth(text);
-        
+
         Color color = getProgressColor();
-        
+
         mc.fontRendererObj.drawString(
-            text,
-            (float) sr.getScaledWidth() / 2.0F / scale - (float) width / 2.0F,
-            (float) sr.getScaledHeight() / 5.0F * 2.0F / scale,
-            color.getRGB() & 16777215 | -1090519040,
-            true
+                text,
+                (float) sr.getScaledWidth() / 2.0F / scale - (float) width / 2.0F,
+                (float) sr.getScaledHeight() / 5.0F * 2.0F / scale,
+                color.getRGB() & 16777215 | -1090519040,
+                true
         );
-        
+
         GlStateManager.disableBlend();
         GlStateManager.enableDepth();
         GL11.glPopMatrix();
@@ -258,15 +258,15 @@ public class AutoBlockIn extends Module {
     private int findBestBlockSlot() {
         int bestSlot = -1;
         int bestScore = Integer.MAX_VALUE;
-        
+
         for (int slot = 0; slot <= 8; slot++) {
             ItemStack stack = mc.thePlayer.inventory.getStackInSlot(slot);
             if (stack == null || stack.stackSize == 0) continue;
-            
+
             if (stack.getItem() instanceof ItemBlock) {
                 Block block = ((ItemBlock) stack.getItem()).getBlock();
                 String blockName = block.getUnlocalizedName().replace("tile.", "");
-                
+
                 Integer score = BLOCK_SCORE.get(blockName);
                 if (score != null && score < bestScore) {
                     bestScore = score;
@@ -275,7 +275,7 @@ public class AutoBlockIn extends Module {
                 }
             }
         }
-        
+
         return bestSlot;
     }
 
@@ -313,7 +313,7 @@ public class AutoBlockIn extends Module {
                     double dx = (x + 0.5) - eye.xCoord;
                     double dy = (y + 0.5) - eye.yCoord;
                     double dz = (z + 0.5) - eye.zCoord;
-                    if (dx*dx + dy*dy + dz*dz > rp12) continue;
+                    if (dx * dx + dy * dy + dz * dz > rp12) continue;
 
                     double d2 = dist2PointAABB(eye, x, y, z);
                     if (d2 > reachSq) continue;
@@ -416,30 +416,32 @@ public class AutoBlockIn extends Module {
         // Try all 6 faces of support block
         for (EnumFacing facing : EnumFacing.values()) {
             BlockPos placementPos = supportBlock.offset(facing);
-            
+
             // Check if placement would be at target
             if (!placementPos.equals(targetPos)) continue;
-            
+
             // Generate candidate hit points on this face
             int n = (int) Math.round(1 / STEP);
-            
+
             for (int r = 0; r <= n; r++) {
                 double v = r * STEP + (Math.random() * JIT * 2 - JIT);
-                if (v < 0) v = 0; else if (v > 1) v = 1;
-                
+                if (v < 0) v = 0;
+                else if (v > 1) v = 1;
+
                 for (int c = 0; c <= n; c++) {
                     double u = c * STEP + (Math.random() * JIT * 2 - JIT);
-                    if (u < 0) u = 0; else if (u > 1) u = 1;
-                    
+                    if (u < 0) u = 0;
+                    else if (u > 1) u = 1;
+
                     Vec3 hitPos = getHitPosOnFace(supportBlock, facing, u, v);
                     float[] rot = getRotationsWrapped(eye, hitPos.xCoord, hitPos.yCoord, hitPos.zCoord);
-                    
+
                     MovingObjectPosition mop = rayTraceBlock(rot[0], rot[1], reach);
-                    if (mop != null 
+                    if (mop != null
                             && mop.typeOfHit == MovingObjectPosition.MovingObjectType.BLOCK
                             && mop.getBlockPos().equals(supportBlock)
                             && mop.sideHit == facing) {
-                        
+
                         targetBlock = supportBlock;
                         targetFacing = facing;
                         targetHitVec = mop.hitVec;
@@ -450,27 +452,27 @@ public class AutoBlockIn extends Module {
                 }
             }
         }
-        
+
         return false;
     }
 
     private void sidesAim(Vec3 eye, double reach, BlockPos feetPos) {
         List<BlockPos> goals = new ArrayList<>();
-        
+
         for (int[] d : DIRS) {
             BlockPos headPos = feetPos.add(d[0], 1, d[2]);
             if (isAir(headPos)) {
                 goals.add(headPos);
             }
         }
-        
+
         for (int[] d : DIRS) {
             BlockPos feetGoal = feetPos.add(d[0], 0, d[2]);
             if (isAir(feetGoal)) {
                 goals.add(feetGoal);
             }
         }
-        
+
         findBestForGoals(goals, eye, reach);
     }
 
@@ -478,31 +480,33 @@ public class AutoBlockIn extends Module {
         for (BlockPos goal : goals) {
             for (EnumFacing facing : EnumFacing.values()) {
                 BlockPos support = goal.offset(facing);
-                
+
                 if (isAir(support)) continue;
-                
+
                 Vec3 center = new Vec3(support.getX() + 0.5, support.getY() + 0.5, support.getZ() + 0.5);
                 if (eye.distanceTo(center) > reach) continue;
-                
+
                 // Try placement
                 int n = (int) Math.round(1 / STEP);
                 for (int r = 0; r <= n; r++) {
                     double v = r * STEP + (Math.random() * JIT * 2 - JIT);
-                    if (v < 0) v = 0; else if (v > 1) v = 1;
-                    
+                    if (v < 0) v = 0;
+                    else if (v > 1) v = 1;
+
                     for (int c = 0; c <= n; c++) {
                         double u = c * STEP + (Math.random() * JIT * 2 - JIT);
-                        if (u < 0) u = 0; else if (u > 1) u = 1;
-                        
+                        if (u < 0) u = 0;
+                        else if (u > 1) u = 1;
+
                         Vec3 hitPos = getHitPosOnFace(support, facing.getOpposite(), u, v);
                         float[] rot = getRotationsWrapped(eye, hitPos.xCoord, hitPos.yCoord, hitPos.zCoord);
-                        
+
                         MovingObjectPosition mop = rayTraceBlock(rot[0], rot[1], reach);
-                        if (mop != null 
+                        if (mop != null
                                 && mop.typeOfHit == MovingObjectPosition.MovingObjectType.BLOCK
                                 && mop.getBlockPos().equals(support)
                                 && mop.sideHit == facing.getOpposite()) {
-                            
+
                             targetBlock = support;
                             targetFacing = facing.getOpposite();
                             targetHitVec = mop.hitVec;
@@ -520,7 +524,7 @@ public class AutoBlockIn extends Module {
         double x = block.getX() + 0.5;
         double y = block.getY() + 0.5;
         double z = block.getZ() + 0.5;
-        
+
         switch (face) {
             case DOWN:
                 y = block.getY() + INSET;
@@ -553,31 +557,31 @@ public class AutoBlockIn extends Module {
                 y = block.getY() + v;
                 break;
         }
-        
+
         return new Vec3(x, y, z);
     }
 
     private boolean isAir(BlockPos pos) {
         Block block = mc.theWorld.getBlockState(pos).getBlock();
-        return block == Blocks.air 
-            || block == Blocks.water 
-            || block == Blocks.flowing_water
-            || block == Blocks.lava
-            || block == Blocks.flowing_lava
-            || block == Blocks.fire;
+        return block == Blocks.air
+                || block == Blocks.water
+                || block == Blocks.flowing_water
+                || block == Blocks.lava
+                || block == Blocks.flowing_lava
+                || block == Blocks.fire;
     }
 
     private void updateProgress() {
         Vec3 playerPos = mc.thePlayer.getPositionVector();
         BlockPos feetPos = new BlockPos(playerPos.xCoord, playerPos.yCoord, playerPos.zCoord);
-        
+
         int filled = 0;
         int total = 9;
-        
+
         if (!isAir(feetPos.up(2))) {
             filled++;
         }
-        
+
         for (int[] d : DIRS) {
             if (!isAir(feetPos.add(d[0], 0, d[2]))) {
                 filled++;
@@ -586,7 +590,7 @@ public class AutoBlockIn extends Module {
                 filled++;
             }
         }
-        
+
         progress = (float) filled / (float) total;
     }
 
@@ -603,14 +607,14 @@ public class AutoBlockIn extends Module {
     private MovingObjectPosition rayTraceBlock(float yaw, float pitch, double range) {
         float yawRad = (float) Math.toRadians(yaw);
         float pitchRad = (float) Math.toRadians(pitch);
-        
+
         double x = -Math.sin(yawRad) * Math.cos(pitchRad);
         double y = -Math.sin(pitchRad);
         double z = Math.cos(yawRad) * Math.cos(pitchRad);
-        
+
         Vec3 start = mc.thePlayer.getPositionEyes(1.0f);
         Vec3 end = start.addVector(x * range, y * range, z * range);
-        
+
         return mc.theWorld.rayTraceBlocks(start, end);
     }
 
@@ -624,16 +628,16 @@ public class AutoBlockIn extends Module {
         double minX = x, maxX = x + 1;
         double minY = y, maxY = y + 1;
         double minZ = z, maxZ = z + 1;
-        
+
         double cx = clamp(p.xCoord, minX, maxX);
         double cy = clamp(p.yCoord, minY, maxY);
         double cz = clamp(p.zCoord, minZ, maxZ);
-        
+
         double dx = p.xCoord - cx;
         double dy = p.yCoord - cy;
         double dz = p.zCoord - cz;
-        
-        return dx*dx + dy*dy + dz*dz;
+
+        return dx * dx + dy * dy + dz * dz;
     }
 
     private double clamp(double v, double lo, double hi) {
@@ -644,13 +648,13 @@ public class AutoBlockIn extends Module {
         double dx = tx - eye.xCoord;
         double dy = ty - eye.yCoord;
         double dz = tz - eye.zCoord;
-        double hd = Math.sqrt(dx*dx + dz*dz);
-        
+        double hd = Math.sqrt(dx * dx + dz * dz);
+
         float yaw = (float) Math.toDegrees(Math.atan2(dz, dx)) - 90.0f;
         yaw = normYaw(yaw);
-        
+
         float pitch = (float) Math.toDegrees(-Math.atan2(dy, hd));
-        
+
         return new float[]{yaw, pitch};
     }
 
@@ -666,7 +670,7 @@ public class AutoBlockIn extends Module {
     private static class BlockData {
         BlockPos pos;
         double distance;
-        
+
         BlockData(BlockPos pos, double distance) {
             this.pos = pos;
             this.distance = distance;
